+++ src/video/vid_hercules.c
@@ -47,8 +47,8 @@ recalc_timings(hercules_t *dev)
     _dispontime *= HERCCONST;
     _dispofftime *= HERCCONST;
 
-    dev->dispontime  = (uint64_t) (_dispontime);
-    dev->dispofftime = (uint64_t) (_dispofftime);
+    dev->dispontime  = (uint64_t) (int64_t) (_dispontime);
+    dev->dispofftime = (uint64_t) (int64_t) (_dispofftime);
 }
 
 static uint8_t crtcmask[32] = {
