+++ src/video/vid_hercules_incolor.c
@@ -196,8 +196,8 @@ recalc_timings(incolor_t *dev)
     _dispontime *= HERCCONST;
     _dispofftime *= HERCCONST;
 
-    dev->dispontime  = (uint64_t) (_dispontime);
-    dev->dispofftime = (uint64_t) (_dispofftime);
+    dev->dispontime  = (uint64_t) (int64_t) (_dispontime);
+    dev->dispofftime = (uint64_t) (int64_t) (_dispofftime);
 }
 
 static void
